feature: product upload renewal
